requirements.txt
# git+https://github.com/facebookresearch/detectron2.git
requests
pillow
opencv-python
colorama